refactor: 테스트도 동일하게 패키지 수정
