+++ src/test/java/kitchenpos/config/RepositoryTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.config;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
