+++ src/test/java/kitchenpos/config/ServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.config;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
