+++ src/test/java/kitchenpos/menu/repository/MenuGroupRepositoryTest.java
@@ -1,10 +1,10 @@
-package kitchenpos.repository;
+package kitchenpos.menu.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.repository.MenuGroupRepository;
+import kitchenpos.config.RepositoryTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
