+++ src/test/java/kitchenpos/menu/repository/MenuProductRepositoryTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.repository;
+package kitchenpos.menu.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,11 +6,9 @@ import java.util.List;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.repository.MenuGroupRepository;
-import kitchenpos.menu.repository.MenuProductRepository;
-import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.config.RepositoryTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
