+++ src/test/java/kitchenpos/menu/repository/MenuRepositoryTest.java
@@ -1,12 +1,11 @@
-package kitchenpos.repository;
+package kitchenpos.menu.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.repository.MenuGroupRepository;
-import kitchenpos.menu.repository.MenuRepository;
+import kitchenpos.config.RepositoryTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
