+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,13 +1,13 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
+import kitchenpos.config.ServiceTest;
 import kitchenpos.product.application.request.ProductCreateRequest;
 import kitchenpos.product.application.response.ProductResponse;
-import kitchenpos.product.application.ProductService;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
