+++ src/test/java/kitchenpos/product/repository/ProductRepositoryTest.java
@@ -1,10 +1,10 @@
-package kitchenpos.repository;
+package kitchenpos.product.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.config.RepositoryTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
