+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -1,16 +1,16 @@
-package kitchenpos.application;
+package kitchenpos.table.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
+import kitchenpos.config.ServiceTest;
 import kitchenpos.table.application.request.OrderTableCreateRequest;
 import kitchenpos.table.application.request.OrderTableUpdateRequest;
 import kitchenpos.table.application.response.OrderTableResponse;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.application.TableService;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.order.repository.OrderRepository;
